+++ documentation/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ### DDT 0.11.1+
  * Fixed: Some editor actions (such as Open Definition) are enabled even when no IDE editor is active.
+ * Added editor Go To Matching Bracket action (shortcut: Ctrl+Shift+P)
 
 ### DDT 0.11.1
  * Fixed: Either Assertion failure or incorrect results when invoking Content Assist on unsaved editor files. (Fixes #99, #97)